feat: Implement thread-safe target deletion and fix in-memory table disposal
Target Verb Integration Tests (46 tests, 100% passing):
- Created comprehensive test suite in tests/integration/test_cases/target_verbs.yaml
- Tests cover all target operations: get, set, clear
- Edge cases: parameter validation, nested tables, lifecycle, stress tests
- Integration tests with other verbs (sizeof, typeof, defined, table.assign)
- Headless-specific behavior tests (no implicit target fallback)

Thread-Safe Target Deletion Implementation:
- Added langunsettarget() in Common/source/langverbs.c
- Clears target as side-effect when deleting a variable set as target
- Prevents crashes when deleting target objects in multi-threaded environments
- Calls langunsettarget() from langdeletefunc() before deletion
- Legacy behavior: deleting target variable clears target (no error)

Critical Bug Fix - In-Memory Table Disposal Crash:
- Fixed segfault in Common/source/langexternal.c:langexternaldisposevariable()
- Root cause: dbpushreleasestack() called with NULL databasedata for in-memory tables
- Tables created with lang.new() have hdatabase=NULL (no database file)
- Added NULL check before dbpushreleasestack() call (line 2426)
- Prevents NULL pointer dereference when disposing in-memory external variables

Infrastructure Support (for future multi-thread work):
- Exported visitprocessthreads() from Common/source/process.c
- Added processinternal.h declarations for thread iteration
- Created helper functions: findouterlocaltable(), unsettargetvisitor()
- Foundation for iterating all threads to clear matching targets

Test Results:
- All 46 target verb integration tests passing
- Delete lifecycle test now works: create → set target → delete → target cleared
- In-memory table deletion no longer crashes

+++ Common/headers/processinternal.h
@@ -269,8 +269,8 @@ extern void copythreadglobals (hdlthreadglobals);
 extern void swapinthreadglobals (hdlthreadglobals);
 
 /* Thread visitor callback type for visitprocessthreads */
-typedef pascal boolean (*threadvisitcallback) (hdlthreadglobals hthread, long refcon);
+typedef pascal boolean (*threadvisitcallback) (hdlthreadglobals hthread, int32_t refcon);
 
-extern boolean visitprocessthreads (threadvisitcallback visit, long refcon);
+extern boolean visitprocessthreads (threadvisitcallback visit, int32_t refcon);
 
 #endif /* processinternalinclude */
